Rename component directories so it matches the component name and titlecase
